hud_django_v1.0/template_django/requirements.txt
asgiref==3.7.2
Django==5.0.6
django-filter==24.2
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
gunicorn==22.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
whitenoise==6.7.0
